refactor(k8s): extract warm admission identity into its own module
`graph_warmer.rs` came in 106 bytes over the 50KB file-size guard once the
identity helpers landed. Rather than reach for an oversize marker, move the
cohesive set — work id, deterministic Job name, and the label stamping that
carries both onto the manifest — into `graph_warmer_identity`, alongside the
budget reasoning that ties all three to Kubernetes label validation.

The regression suite moves with the code it covers. No behaviour change.
