+++ server/crates/djinn-k8s/src/lib.rs
@@ -8,6 +8,7 @@
 pub mod config;
 pub mod env_config;
 pub mod graph_warmer;
+pub mod graph_warmer_identity;
 pub mod infra_death_log_tail;
 pub mod job;
 pub mod label_value;
@@ -28,8 +29,9 @@ pub use graph_warmer::{
     K8sGraphWarmer, KubeClientDispatcher, KubeClientJobWatcher, KubeClientWarmJobLister,
     NoopJobWatcher, NoopWarmJobLister, WarmAdmission, WarmAdmissionError, WarmAdmissionPermit,
     WarmAdmissionRequest, WarmAdmissionTransition, WarmCompletionSink, WarmJobDispatcher,
-    WarmJobLister, WarmJobManifest, WarmJobWatcher, WarmTerminalOutcome, warm_work_id,
+    WarmJobLister, WarmJobManifest, WarmJobWatcher, WarmTerminalOutcome,
 };
+pub use graph_warmer_identity::warm_work_id;
 pub use runtime::KubernetesRuntime;
 pub use token_review::TokenReviewer;
 pub use warm_job::build_warm_job;
